+++ ErsatzTV.Core/Interfaces/Metadata/ILocalFileSystem.cs
@@ -14,6 +14,7 @@ namespace ErsatzTV.Core.Interfaces.Metadata
         IEnumerable<string> ListSubdirectories(string folder);
         IEnumerable<string> ListFiles(string folder);
         bool FileExists(string path);
+        bool FolderExists(string folder);
         Task<Either<BaseError, Unit>> CopyFile(string source, string destination);
         Unit EmptyFolder(string folder);
     }